Enhance wheel install command to allow the user to specify requirements.
The wheel install command now takes requirement specifications (as per
pkg_resources, for example "foo" to install the best version of foo available,
or "bar==1.0" to install bar version 1.0) as well as exact filenames to
install.

Also added a --list option to wheel install, to list what will be installed
(but not install it).

Internal change - WheelFile objects are now comparable, one wheelfile is
greater than another if it is "better" - where the wheels are for different
packages, this just means the package name sorts later, but if they are for
the same package, the later version is better, and more compatible is better
within versions. So, to get the most appropriate wheel to install from a list,
just use max(wheel_list).